Effect aggregation
if a skill or subskill has a main_skill (aka an actual usable skill) the on_hit_effects of the upgrades are forwarded to this skill.
if a main_skill is not usable (aka it is not the root skill) it will only get the effects in the skill and not the upgrades.

self_effects are all aggregated into one list and then applied to the caster